Bubble-wrapped.
