+++ Rakefile
@@ -1,6 +1,6 @@
 require 'rake'
 require 'rake/testtask'
-require 'rake/rdoctask'
+require 'rdoc/task'
 
 # TODO - want other tests/tasks run by default? Add them to the list
 # task :default => [:spec, :features]
