+++ lib/random_data/dates.rb
@@ -1,5 +1,3 @@
-require 'date'
-
 module RandomData
   
   # Defines methods for random date generation
